Update 5xx template
